Reordered last paragraphs
Other thing I want to inspire is that even if I want a HTTPS internet sometimes we are just over engineering it, like encrypting again in the transport something that was encrypted in the application. I will try to use HTTP to reduce server costs on P2P signaling in order to make it cheaper to deploy and allowing it to be more decentralized.